network: refactor (deduplicate, remove unnecessary code) subnet-related entities across records & commons sub-packages
