+++ network/network.go
@@ -2,10 +2,10 @@ package network
 
 import (
 	"context"
+	"github.com/ssvlabs/ssv/network/commons"
 	"io"
 
 	"github.com/libp2p/go-libp2p/core/peer"
-	"github.com/ssvlabs/ssv/network/records"
 	protocolp2p "github.com/ssvlabs/ssv/protocol/v2/p2p"
 	"go.uber.org/zap"
 )
@@ -48,9 +48,9 @@ type P2PNetwork interface {
 	// UpdateScoreParams will update the scoring parameters of GossipSub
 	UpdateScoreParams(logger *zap.Logger)
 	// ActiveSubnets returns active subnets
-	ActiveSubnets() records.Subnets
+	ActiveSubnets() commons.Subnets
 	// FixedSubnets returns fixed subnets
-	FixedSubnets() records.Subnets
+	FixedSubnets() commons.Subnets
 
 	// used for tests and api
 	PeersByTopic() ([]peer.ID, map[string][]peer.ID)
