+++ network/p2p/p2p_reporter.go
@@ -3,10 +3,8 @@ package p2pv1
 import (
 	"math"
 
-	"github.com/ssvlabs/ssv/logging/fields"
-	"github.com/ssvlabs/ssv/network/commons"
-
 	spectypes "github.com/ssvlabs/ssv-spec/types"
+	"github.com/ssvlabs/ssv/logging/fields"
 	"go.uber.org/zap"
 
 	ssvpeers "github.com/ssvlabs/ssv/network/peers"
@@ -19,7 +17,7 @@ func (n *p2pNetwork) ReportValidation(logger *zap.Logger, msg *spectypes.SSVMess
 	if !n.isReady() {
 		return
 	}
-	data, err := commons.EncodeNetworkMsg(msg)
+	data, err := msg.Encode()
 	if err != nil {
 		logger.Warn("could not encode message", zap.Error(err))
 		return
