+++ network/peers/connections/helpers_test.go
@@ -49,7 +49,7 @@ func getTestingData(t *testing.T) TestData {
 			NodeVersion:   "some-node-version",
 			ExecutionNode: "some-execution-node",
 			ConsensusNode: "some-consensus-node",
-			Subnets:       records.AllSubnets,
+			Subnets:       commons.AllSubnets,
 		},
 	}
 
@@ -60,7 +60,7 @@ func getTestingData(t *testing.T) TestData {
 				NodeVersion:   "test-node-version",
 				ExecutionNode: "test-execution-node",
 				ConsensusNode: "test-consensus-node",
-				Subnets:       records.AllSubnets,
+				Subnets:       commons.AllSubnets,
 			},
 		},
 		MockSelfSealed: []byte("something"),
@@ -93,7 +93,7 @@ func getTestingData(t *testing.T) TestData {
 		ctx:        context.Background(),
 		nodeInfos:  nii,
 		peerInfos:  ns,
-		subnetsIdx: peers.NewSubnetsIndex(commons.Subnets()),
+		subnetsIdx: peers.NewSubnetsIndex(commons.SubnetsCount),
 		ids:        ids,
 		net:        net,
 		streams:    sc,
