+++ network/peers/index.go
@@ -1,6 +1,7 @@
 package peers
 
 import (
+	"github.com/ssvlabs/ssv/network/commons"
 	"io"
 
 	"github.com/libp2p/go-libp2p/core/network"
@@ -99,13 +100,13 @@ type SubnetsStats struct {
 // it keeps track of subnets but doesn't mind regards actual connections that we have.
 type SubnetsIndex interface {
 	// UpdatePeerSubnets updates the given peer's subnets
-	UpdatePeerSubnets(id peer.ID, s records.Subnets) bool
+	UpdatePeerSubnets(id peer.ID, s commons.Subnets) bool
 
 	// GetSubnetPeers returns peers that are interested in the given subnet
 	GetSubnetPeers(s int) []peer.ID
 
 	// GetPeerSubnets returns subnets of the given peer
-	GetPeerSubnets(id peer.ID) records.Subnets
+	GetPeerSubnets(id peer.ID) commons.Subnets
 
 	// GetSubnetsStats collects and returns subnets stats
 	GetSubnetsStats() *SubnetsStats
