+++ operator/validator/metadata/syncer_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"crypto/rand"
 	"fmt"
+	"github.com/ssvlabs/ssv/network/commons"
 	"testing"
 	"time"
 
@@ -16,7 +17,6 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/ssvlabs/ssv/logging"
-	"github.com/ssvlabs/ssv/network/records"
 	"github.com/ssvlabs/ssv/networkconfig"
 	"github.com/ssvlabs/ssv/protocol/v2/blockchain/beacon"
 	ssvtypes "github.com/ssvlabs/ssv/protocol/v2/types"
@@ -95,7 +95,7 @@ func TestUpdateValidatorMetadata(t *testing.T) {
 				return result, nil
 			}).AnyTimes()
 
-			noSubnets, err := records.Subnets{}.FromString("0x00000000000000000000000000000000")
+			noSubnets, err := commons.Subnets{}.FromString("0x00000000000000000000000000000000")
 			require.NoError(t, err)
 
 			syncer := NewSyncer(logger, sharesStorage, validatorStore, networkconfig.TestNetwork.Beacon, beaconNode, noSubnets)
@@ -663,7 +663,7 @@ func TestWithUpdateInterval(t *testing.T) {
 	// Define the interval we want to set
 	interval := testSyncInterval * 2
 
-	noSubnets, err := records.Subnets{}.FromString("0x00000000000000000000000000000000")
+	noSubnets, err := commons.Subnets{}.FromString("0x00000000000000000000000000000000")
 	require.NoError(t, err)
 
 	// Create an Syncer with the WithSyncInterval option
